changed double to BigDecimal, unhardcoded role

+++ src/main/java/app/diary/ProductAddedToDiary.java
@@ -1,14 +1,13 @@
 package app.diary;
 
-import app.diary.Diary;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -22,11 +21,11 @@ import lombok.NoArgsConstructor;
 
     private String productId;
     private String productName;
-    private double kcal;
-    private double protein;
-    private double carbohydrates;
-    private double fat;
-    private double fiber;
+    private BigDecimal kcal;
+    private BigDecimal protein;
+    private BigDecimal carbohydrates;
+    private BigDecimal fat;
+    private BigDecimal fiber;
     private String image;
     private String measureLabel;
     private double quantity;
